Refactor Enforcer Admin API dtos: moved into a different package to keep them isolated
