+++ enforcer-parent/enforcer/src/main/java/org/wso2/choreo/connect/enforcer/admin/handlers/APIRequestHandler.java
@@ -23,15 +23,15 @@ import org.wso2.choreo.connect.enforcer.admin.AdminUtils;
 import org.wso2.choreo.connect.enforcer.api.API;
 import org.wso2.choreo.connect.enforcer.api.APIFactory;
 import org.wso2.choreo.connect.enforcer.constants.AdminConstants;
-import org.wso2.choreo.connect.enforcer.models.APIInfo;
-import org.wso2.choreo.connect.enforcer.models.APIList;
 import org.wso2.choreo.connect.enforcer.models.Application;
-import org.wso2.choreo.connect.enforcer.models.ApplicationInfo;
 import org.wso2.choreo.connect.enforcer.models.ApplicationKeyMapping;
-import org.wso2.choreo.connect.enforcer.models.BasicAPIInfo;
 import org.wso2.choreo.connect.enforcer.models.ResponsePayload;
 import org.wso2.choreo.connect.enforcer.models.Subscription;
-import org.wso2.choreo.connect.enforcer.models.SubscriptionInfo;
+import org.wso2.choreo.connect.enforcer.models.admin.APIInfo;
+import org.wso2.choreo.connect.enforcer.models.admin.APIList;
+import org.wso2.choreo.connect.enforcer.models.admin.ApplicationInfo;
+import org.wso2.choreo.connect.enforcer.models.admin.BasicAPIInfo;
+import org.wso2.choreo.connect.enforcer.models.admin.SubscriptionInfo;
 
 import java.util.ArrayList;
 import java.util.List;
