+++ enforcer-parent/enforcer/src/main/java/org/wso2/choreo/connect/enforcer/models/admin/APIInfo.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2022, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ * Copyright (c) 2021, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
  *
  * WSO2 Inc. licenses this file to you under the Apache License,
  * Version 2.0 (the "License"); you may not use this file except
@@ -15,23 +15,24 @@
  * specific language governing permissions and limitations
  * under the License.
  */
+package org.wso2.choreo.connect.enforcer.models.admin;
 
-package org.wso2.choreo.connect.enforcer.models;
+import java.util.List;
 
 /**
- * Entity for representing API related information for Admin REST API within enforcer.
+ * Model class for API Information
  */
-public class BasicAPIInfo {
-
+public class APIInfo {
     private String provider = null;
     private String name = null;
     private String version = null;
     private String context = null;
-    private String policy = null;
+    private String tier = null;
     private String apiType = null;
     private boolean isDefaultVersion = false;
     private String apiUUID = null;
     private String lcState = null;
+    private List<SubscriptionInfo> subscriptions = null;
 
     public String getProvider() {
         return provider;
@@ -49,6 +50,14 @@ public class BasicAPIInfo {
         this.name = name;
     }
 
+    public String getTier() {
+        return tier;
+    }
+
+    public void setTier(String tier) {
+        this.tier = tier;
+    }
+
     public String getVersion() {
         return version;
     }
@@ -63,14 +72,6 @@ public class BasicAPIInfo {
 
     public void setContext(String context) {
         this.context = context;
-    }
-
-    public String getPolicy() {
-        return policy;
-    }
-
-    public void setPolicy(String policy) {
-        this.policy = policy;
     }
 
     public String getApiType() {
@@ -103,5 +104,13 @@ public class BasicAPIInfo {
 
     public void setLcState(String lcState) {
         this.lcState = lcState;
+    }
+
+    public List<SubscriptionInfo> getSubscriptions() {
+        return subscriptions;
+    }
+
+    public void setSubscriptions(List<SubscriptionInfo> subscriptions) {
+        this.subscriptions = subscriptions;
     }
 }
