+++ enforcer-parent/enforcer/src/main/java/org/wso2/choreo/connect/enforcer/models/admin/ApplicationInfo.java
@@ -15,7 +15,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.wso2.choreo.connect.enforcer.models;
+package org.wso2.choreo.connect.enforcer.models.admin;
 
 import java.util.ArrayList;
 import java.util.List;
