+++ enforcer-parent/enforcer/src/main/java/org/wso2/choreo/connect/enforcer/models/admin/ApplicationInfoList.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2020, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ * Copyright (c) 2021, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
  *
  * WSO2 Inc. licenses this file to you under the Apache License,
  * Version 2.0 (the "License"); you may not use this file except
@@ -15,37 +15,32 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-
-package org.wso2.choreo.connect.enforcer.models;
+package org.wso2.choreo.connect.enforcer.models.admin;
 
 import java.util.ArrayList;
 import java.util.List;
 
 /**
- * Holds details about list of APIs.
+ * Model class for Application Information list
  */
-public class APIList {
+public class ApplicationInfoList {
 
     private Integer count = null;
-    private List<BasicAPIInfo> list = new ArrayList<>();
+    private List<ApplicationInfo> list = new ArrayList<>();
 
     public Integer getCount() {
-
         return count;
     }
 
     public void setCount(Integer count) {
-
         this.count = count;
     }
 
-    public List<BasicAPIInfo> getList() {
-
+    public List<ApplicationInfo> getList() {
         return list;
     }
 
-    public void setList(List<BasicAPIInfo> list) {
-
+    public void setList(List<ApplicationInfo> list) {
         this.list = list;
     }
 }
